refactor: 메소드 AuthAcceptanceTest 로 내림
